+++ src/app/(main)/profile/[userId]/page.tsx
@@ -1,6 +1,7 @@
 import ProfileByIdDetail from './components/ProfileByIdDetail.client';
 import ProfileByIdHeader from './components/ProfileByIdHeader';
 import { Keys, getProfileById } from '@/apis/profile';
+import { PageAnimation } from '@/components/PageAnimation';
 import { HydrationProvider } from '@/components/Provider';
 import { Suspense } from 'react';
 
@@ -15,13 +16,15 @@ export default function page({ params }: PageProps) {
 
   return (
     <Suspense>
-      <HydrationProvider
-        queryFn={() => getProfileById(userId)}
-        queryKey={Keys.getProfileById(userId)}
-      >
-        <ProfileByIdHeader />
-        <ProfileByIdDetail />
-      </HydrationProvider>
+      <PageAnimation className="bg-sub">
+        <HydrationProvider
+          queryFn={() => getProfileById(userId)}
+          queryKey={Keys.getProfileById(userId)}
+        >
+          <ProfileByIdHeader />
+          <ProfileByIdDetail />
+        </HydrationProvider>
+      </PageAnimation>
     </Suspense>
   );
 }
